backend/requirements.txt
# Core application
fastapi
uvicorn

# Pydantic and settings
pydantic
pydantic-settings
python-dotenv

# Langchain for AI orchestration
langchain
langchain-core
langchain-community
langchain-text-splitters
SQLAlchemy

# Database connector
oracledb

# Other essential dependencies
anyio
click
h11
httpx
idna
Jinja2
MarkupSafe
python-dateutil
sniffio
starlette
typing_extensions
PyYAML
requests
numpy
packaging

# Google Gemini AI
google-generativeai

# SQL parsing and validation
sqlparse
